feat(serial,eeprom): add option to disable
Prior to this commit, the EEPROM API and serial handler were both included.
However, due to lack of resources on some architectures (e.g., insufficient
flash memory on `atmega8`), it may be useful to optionally exclude either the
EEPROM API or the serial handler from the build.

In this commit, use the `NO_EEPROM` and `NO_SERIAL` preprocessor definitions to
disable the corresponding code. By default, these are **not** set.